This commit fixes alignment of the copied to table function.
Since reporting has been fixed a fucntion has been added to copy the report
as a table so that it be pasted in an email and report out attendance without
having to attach or upload a file.

Also, the function to export to a .csv has been added just in case ;).